chore(eslint): stop extending from unexisting plugin - prettier
Currently, the this project's ESLint config extends from
`"plugin:prettier/recommended"`. However, the `eslint-plugin-prettier`
package is not a dependency of the project and running ESLint throws an error. This PR removes prettier from the ESLint configuration. Note that users can still use `prettier` for formatting by running `npx prettier .` or using an IDE, such as VSCode with the `prettier` extension.